safari-600.7-branch/LayoutTests/fast/css/image-set-parsing-invalid-expected.txt

Test the parsing of the -webkit-image-set function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Too many url parameters : url(#a #b)
PASS cssRule is null

No x : url('#a') 1
PASS cssRule is null

No comma : url('#a') 1x url('#b') 2x
PASS cssRule is null

Too many scale factor parameters : url('#a') 1x 2x
PASS cssRule is null

Scale factor is 0 : url('#a') 0x
PASS cssRule is null

No url function : '#a' 1x
PASS cssRule is null
PASS successfullyParsed is true

TEST COMPLETE
